Address pull request comments to improve the code
